requirements.txt
asgiref==3.8.1
asttokens==3.0.0
boto3==1.35.88
botocore==1.35.88
certifi==2024.12.14
colorama==0.4.6
Django==5.1
django-cors-headers==4.6.0
django-extensions==3.2.3
django-rest-knox==5.0.1
djangorestframework==3.15.2
executing==2.1.0
icecream==2.1.3
jmespath==1.0.1
pillow==11.0.0
psycopg2==2.9.10
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.4
sentry-sdk==2.19.2
six==1.17.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.3.0
